Move Label to trees/generic/def.rs
