+++ lang/lowering/src/ctx.rs
@@ -83,7 +83,7 @@ impl Ctx {
         &mut self,
         user_name: Option<Ident>,
         info: &Span,
-    ) -> Result<Label, LoweringError> {
+    ) -> Result<ust::Label, LoweringError> {
         if let Some(user_name) = &user_name {
             match Context::lookup(self, user_name) {
                 Some(Elem::Decl(_)) => {
@@ -110,7 +110,7 @@ impl Ctx {
         }
         let id = self.next_label_id;
         self.next_label_id += 1;
-        Ok(Label { id, user_name })
+        Ok(ust::Label { id, user_name })
     }
 
     /// Next De Bruijn level to be assigned
