+++ lang/syntax/src/common/mod.rs
@@ -1,5 +1,3 @@
-use std::fmt;
-
 use codespan::Span;
 
 mod de_bruijn;
@@ -8,7 +6,6 @@ mod named;
 mod subst;
 
 pub use de_bruijn::*;
-use derivative::Derivative;
 pub use equiv::*;
 pub use named::*;
 pub use subst::*;
@@ -23,24 +20,5 @@ pub trait HasSpan {
 impl HasSpan for Option<Span> {
     fn span(&self) -> Option<Span> {
         *self
-    }
-}
-
-#[derive(Debug, Clone, Derivative)]
-#[derivative(Eq, PartialEq, Hash)]
-pub struct Label {
-    /// A machine-generated, unique id
-    pub id: usize,
-    /// A user-annotated name
-    #[derivative(PartialEq = "ignore", Hash = "ignore")]
-    pub user_name: Option<Ident>,
-}
-
-impl fmt::Display for Label {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.user_name {
-            None => Ok(()),
-            Some(user_name) => user_name.fmt(f),
-        }
     }
 }
