+++ lang/syntax/src/trees/generic/def.rs
@@ -47,6 +47,25 @@ where
     }
 }
 
+#[derive(Debug, Clone, Derivative)]
+#[derivative(Eq, PartialEq, Hash)]
+pub struct Label {
+    /// A machine-generated, unique id
+    pub id: usize,
+    /// A user-annotated name
+    #[derivative(PartialEq = "ignore", Hash = "ignore")]
+    pub user_name: Option<Ident>,
+}
+
+impl fmt::Display for Label {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match &self.user_name {
+            None => Ok(()),
+            Some(user_name) => user_name.fmt(f),
+        }
+    }
+}
+
 #[derive(Debug, Clone)]
 pub struct DocComment {
     pub docs: Vec<String>,
