+++ lang/syntax/src/trees/ust/def.rs
@@ -13,6 +13,7 @@ impl generic::Phase for UST {
     type Ctx = ();
 }
 
+pub type Label = generic::Label;
 pub type DocComment = generic::DocComment;
 pub type Attribute = generic::Attribute;
 pub type Prg = generic::Prg<UST>;
